Add deletion confirmation message

+++ kubehandler/manageDeploy.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"log"
 	"strconv"
+	"time"
 
 	"github.com/bwmarrin/discordgo"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -32,21 +33,58 @@ func UpdateDeployment(s *discordgo.Session, m *discordgo.MessageCreate, namespac
 
 // DeleteDeployment function to delete deployments
 func DeleteDeployment(s *discordgo.Session, m *discordgo.MessageCreate, namespace string, deploy string) {
-	deletePolicy := metav1.DeletePropagationForeground
-	err := clientset.AppsV1().Deployments(namespace).Delete(context.TODO(), deploy, metav1.DeleteOptions{
-		PropagationPolicy: &deletePolicy,
-	})
-	if err != nil {
-		log.Printf("Error in deleting deployment : %v", err)
+
+	var confirmVar bool
+
+	confirmMsg, confErr := s.ChannelMessageSend(m.ChannelID, "Please react with ✅ emoji within 1 min to confirm deletion of "+deploy)
+	if confErr != nil {
+		log.Printf("Error in sending message: %v", confErr)
 	}
 
-	msg := &discordgo.MessageEmbed{
-		Author: &discordgo.MessageEmbedAuthor{
-			Name: deploy + " Deleted",
-		},
-		Description: "Successfully deleted deployment " + deploy,
+	time.Sleep(1 * time.Minute)
+
+	updatedMsg, err := s.ChannelMessage(m.ChannelID, confirmMsg.ID)
+	if err != nil {
+		log.Printf("Error in getting message: %v", err)
 	}
-	s.ChannelMessageSendEmbed(m.ChannelID, msg)
+	if len(updatedMsg.Reactions) > 1 {
+		confirmVar = false
+	} else {
+		for _, i := range updatedMsg.Reactions {
+			if i.Emoji.Name == "✅" {
+				confirmVar = true
+			} else {
+				confirmVar = false
+			}
+		}
+	}
+
+	if confirmVar == true {
+		deletePolicy := metav1.DeletePropagationForeground
+		err := clientset.AppsV1().Deployments(namespace).Delete(context.TODO(), deploy, metav1.DeleteOptions{
+			PropagationPolicy: &deletePolicy,
+		})
+		if err != nil {
+			log.Printf("Error in deleting deployment : %v", err)
+		}
+
+		msg := &discordgo.MessageEmbed{
+			Author: &discordgo.MessageEmbedAuthor{
+				Name: deploy + " Deleted",
+			},
+			Description: "Successfully deleted deployment " + deploy,
+		}
+		s.ChannelMessageSendEmbed(m.ChannelID, msg)
+	} else {
+
+		msg := &discordgo.MessageEmbed{
+			Author: &discordgo.MessageEmbedAuthor{
+				Name: deploy + " Deletion Aborted",
+			},
+			Description: "Deletetion aborted for deployment " + deploy,
+		}
+		s.ChannelMessageSendEmbed(m.ChannelID, msg)
+	}
 
 }
 func int32Ptr(i string) *int32 {
